requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
boto3==1.28.27
botocore==1.31.27
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.4.2
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.7.0
pymongo==4.4.1
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
s3transfer==0.6.2
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.4
WTForms==3.0.1
